AutoUpdater:  * Added new certificate, so the AutoUpdate will work also with the new server certificate (which will be active latest on 27.01.2012)  * The AutoUpdate now uses two certificates for seamless updates during server-certificate-change)  * Changed path where the AutoUpdate looks for updates (due to general website update); the old path still works, but will be deactivated in future.
Note: AutoUpdates are possible again for '''all''' versions, due to reverting the SSL-server-certificate to the old one. However, one '''must''' update before 27.01.2012 to to a version >= 4447! After 27.01.2012, there is no way to AutoUpdate a version <4447! In this case one has to manually install a new version (after that autoupdate will work again)


git-svn-id: https://svn.cryptool.org/CrypTool2/trunk@4447 9b0d660e-8ab2-dd11-ad6d-000c29d71a56