New component added: Keccak (SHA-3)
The component enables the usage of Keccak as a hash function with the parameters for SHA3-224, SHA3-256, SHA3-384 and SHA3-512. It is also possible to tweak the parameters such that the component can be used as a one-way function with arbitrary output size (for example as a stream cipher or a prng).
Three templates are added, to show Keccak as hash function, Keccak as stream cipher (the component generates a keystream which is XOR'd on the plaintext) and Keccak as a prng, which is used to generate a random key and a random IV for the PRESENT cipher.


git-svn-id: https://svn.cryptool.org/CrypTool2/trunk@5276 9b0d660e-8ab2-dd11-ad6d-000c29d71a56