Wizard:  * Added workaround for a small problem with #665  * Now the wizard path "Cryptography->Modern->Symmetric->AES" fully works incl. decryption. This has to be adapted for the other 5 paths; refs #757
This refs #665:
Setting one plugin as attribute in the comboBox is required for now as a workaround. Without this attribute, the wizard forgets the settings from the last page. It manifest like this: when going one step backward it shows the right settings in the GUI, but in the background the wizard uses the initial value.


git-svn-id: https://svn.cryptool.org/CrypTool2/trunk@5349 9b0d660e-8ab2-dd11-ad6d-000c29d71a56